Separate search result ad JSON-ld parsing logic.
fix https://github.com/brave/brave-browser/issues/22744
