+++ components/brave_ads/content/browser/search_result_ad/BUILD.gn
@@ -1,5 +1,7 @@
 source_set("search_result_ad") {
   sources = [
+    "search_result_ad_parsing.cc",
+    "search_result_ad_parsing.h",
     "search_result_ad_service.cc",
     "search_result_ad_service.h",
   ]
@@ -9,6 +11,7 @@ source_set("search_result_ad") {
     "//brave/components/brave_ads/browser",
     "//brave/components/brave_ads/common",
     "//brave/components/brave_search/common",
+    "//components/schema_org/common:mojom",
     "//content/public/browser",
     "//mojo/public/cpp/bindings",
     "//services/service_manager/public/cpp",
